Moving test-cases file
