+++ stylelint-rules/lib/_test-cases.js
